sar-5 updated id data type integer to long

+++ src/main/java/com/dft/api/shopify/v202307/model/fullfillment/ShopifyFulfillment.java
@@ -16,8 +16,8 @@ import java.time.LocalDateTime;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class ShopifyFulfillment {
 
-    private Integer id;
-    private Integer orderId;
+    private Long id;
+    private Long orderId;
     private String status;
     private String trackingCompany;
     private String trackingNumber;
